Add a python script to help identify new patterns
This commit intends to add a python script that reads all the tables in the generic_matches and upheno_matches directories. The upheno_matches tsv files represent patterns that are already defined.
The script outputs a table that contains generic_patterns *minus* matching upheno_patterns.
If applied, this commit is related to #31.